backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.1
sqlalchemy>=2.0.30
edge-tts>=6.1.12
httpx>=0.27.0
groq>=0.9.0
slowapi>=0.1.9
python-multipart>=0.0.9
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
pillow>=10.3.0
moviepy>=2.0.0
imageio-ffmpeg>=0.5.1
